requirements.txt
certifi==2025.4.26
distlib==0.3.9
filelock==3.18.0
iniconfig==2.1.0
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pipenv==2025.0.2
platformdirs==4.3.8
pluggy==1.6.0
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
setuptools==80.8.0
six==1.17.0
tzdata==2025.2
virtualenv==20.31.2
